requirements.txt
celery==4.2.1
Django==1.11.13
django-cors-middleware==1.3.1
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
drf-yasg==1.14.0
gunicorn==20.0.4
kafka-python==1.4.3
kombu==4.2.1
psycopg2-binary==2.7.5
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
